requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-huggingface>=0.0.10
openai==1.3.0
faiss-cpu==1.7.4
tiktoken==0.4.0
streamlit>=1.20.0
python-dotenv>=0.21.0
sentence-transformers>=2.5.0
torch>=2.0.0
sentence-transformers>=2.5.0
torch>=2.0.0 